+++ src/surreal.service.ts
@@ -1,0 +1,13 @@
+import { Injectable, OnModuleDestroy } from '@nestjs/common';
+import Surreal from 'surrealdb.js';
+
+@Injectable()
+export class SurrealService extends Surreal implements OnModuleDestroy {
+  constructor() {
+    super();
+  }
+
+  async onModuleDestroy() {
+    await this.close();
+  }
+}
